Move LoginController.cs -> Data/Controller/ Move WeatherForecastController.cs -> Data/Controller/Data
Move Users.cs -> Data/Models
Move WeatherForecast.cs -> Manager

Add JwtTokenReference to the Config
