+++ AuthTest/Data/Models/User.cs
@@ -1,0 +1,8 @@
+﻿namespace AuthTest.Data.Models
+{
+    public class User
+    {
+        public string UserName { get; set; }
+        public string Password { get; set; }
+    }
+}
