--- AuthTest/IJwtAuthenticationManager.cs
@@ -1,9 +1,0 @@
-﻿namespace AuthTest
-{
-    public interface IJwtAuthenticationManager
-    {
-        string Authenticate(string username, string password);
-
-       
-    }
-}
