+++ AuthTest/appsettings.json
@@ -5,5 +5,8 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "Jwt": {
+    "Token": "YourVerySecureTokenWithAllSecurityRequirements"
+  }
 }
